// packages/tinker-mbti-test/src/types.ts
export type Dimension = 'EI' | 'SN' | 'TF' | 'JP'

export type Choice = 'a' | 'b'

export type MBTIType =
  | 'ENFJ'
  | 'ENFP'
  | 'ENTJ'
  | 'ENTP'
  | 'ESFJ'
  | 'ESFP'
  | 'ESTJ'
  | 'ESTP'
  | 'INFJ'
  | 'INFP'
  | 'INTJ'
  | 'INTP'
  | 'ISFJ'
  | 'ISFP'
  | 'ISTJ'
  | 'ISTP'

export type Screen = 'intro' | 'question' | 'result'